feat: Implementation of merge process as described in issue #1
