+++ README.md
@@ -2,7 +2,7 @@
 
 This project is provided by the OpenAPI Initiative as a means to centralize ecosystem information on OpenAPI-related tooling. It leverages open source projects that have gone before to provide a consolidated list of tooling.
 
-This project Kanban board for Tooling can be found here: https://github.com/OAI/Projects/projects/4
+The project Kanban board for Tooling can be found here: https://github.com/OAI/Projects/projects/4
 
 ## Roll Call
 
@@ -27,17 +27,43 @@ Each is expanded upon in the sections below.
 
 The tooling list is built in largely the same format as the majority of projects that have blazed a trail in tooling before (which of course this project takes full advantage of).
 
-In order to bring this together in a sensible way a Gulp-based process has been implemented. Gulp was chosen given the relative ease with which functions can be implemented to massage the data stream and to ensure the build is not closely-coupled to a CI tool.
+In order to bring this together in a sensible way a Gulp-based process has been implemented. Gulp was chosen given the relative ease with which functions can be implemented to massage the data stream and to ensure the build is not closely-coupled to a (commercial) CI tool. The functions themselves are abstracted away from Gulp to enable the build to "lift-and-shift" to a new build tool as required.
 
-Currently only the initial build is implemented. The steps are as follows:
+### Full Build
 
-* Retrieve each tooling source.
+The full build takes the following approach:
+
+* Retrieve each tooling source, including the existing list at [docs/tools.yaml](docs/tools.yaml).
 * Combine source data based on repository name.
 * Normalise property names across sources using simple statistics (Sørensen–Dice, Damerau–Levenshtein distance).
 * Get repository metadata from GitHub.
-* Write to [docs/tools.yaml](docs/tools.yaml)
+* Write to [docs/tools.yaml](docs/tools.yaml).
 
-This process will be amended to include a merge process to capture updates to sources and merge into the existing dataset.
+Currently this build is not scheduled and is [in the backlog](https://github.com/OAI/Tooling/issues/9) to implement.
+
+To run the full build:
+
+```bash
+yarn install
+gulp full
+```
+
+### Metadata Update
+
+The goal of the metadata update is to provide consistent repository metadata without sourcing new tooling:
+
+* Read the existing list at [docs/tools.yaml](docs/tools.yaml).
+* Get repository metadata from GitHub.
+* Write to [docs/tools.yaml](docs/tools.yaml).
+
+Currently this build is not scheduled and is [in the backlog](https://github.com/OAI/Tooling/issues/9) to implement.
+
+To run the metadata build:
+
+```bash
+yarn install # If you haven't done this already
+gulp metadata
+```
 
 ### Website
