--- gulpfile.js/lib/processors/default-processor.js
@@ -1,1 +1,0 @@
-module.exports = async () => [];
