--- gulpfile.js/sourceMetadata.json
@@ -1,21 +1,0 @@
-[
-  {
-    "title": "IMPLEMENTATIONS.md",
-    "url": "https://raw.githubusercontent.com/OAI/OpenAPI-Specification/main/IMPLEMENTATIONS.md",
-    "processor": "lib/processors/openapi-implementations-md-processor.js"
-  },
-  {
-    "title": "https://openapi.tools/",
-    "url": "https://raw.githubusercontent.com/apisyouwonthate/openapi.tools/master/_data/tools.yml",
-    "processor": "lib/processors/openapi-tools-processor.js"
-  },
-  {
-    "title": "openapi3 tags",
-    "processor": "lib/processors/awesome-openapi3-processor.js"
-  },
-  {
-    "title": "local static data",
-    "url": "src/_data/static.yaml",
-    "processor": "lib/processors/default-processor.js"
-  }
-]
