+++ lib/index.js
@@ -1,0 +1,10 @@
+const {
+  getRepositoryMetadata, readSourceData, mergeSources, normaliseSources,
+} = require('./transform');
+
+module.exports = {
+  getRepositoryMetadata,
+  readSourceData,
+  mergeSources,
+  normaliseSources,
+};
