+++ lib/processors/openapi-tools-processor.js
@@ -1,0 +1,13 @@
+const http = require('axios');
+const YAML = require('js-yaml');
+
+const { logger } = require('../util');
+
+module.exports = async (source, url) => {
+  const response = await http.get(url);
+  const data = YAML.load(response.data);
+
+  logger('Number of sources found in openapi.tools', data.length);
+
+  return data.map((tool) => ({ source, ...tool }));
+};
