+++ lib/repo/index.js
@@ -1,0 +1,5 @@
+const getGithubRepositoryMetadata = require('./github');
+
+module.exports = {
+  getGithubRepositoryMetadata,
+};
