+++ package.json
@@ -6,9 +6,10 @@
   "license": "MIT",
   "scripts": {
     "build": "gulp",
-    "coverage": "export NODE_ENV=test && nyc --reporter=html --reporter=text ./node_modules/.bin/_mocha gulpfile.js/test --exit --recursive --timeout 10000",
+    "coverage": "export NODE_ENV=test && nyc --reporter=html --reporter=text ./node_modules/.bin/_mocha gulpfile.js/test/lib/* --exit --recursive --timeout 10000",
+    "coverage:desktop": "yarn run coverage && open coverage/index.html",
     "start": "npx @11ty/eleventy --serve",
-    "test": "export NODE_ENV=test && mocha gulpfile.js/test --exit --recursive --timeout 10000"
+    "test": "export NODE_ENV=test && mocha test --exit --recursive --timeout 10000"
   },
   "dependencies": {
     "@11ty/eleventy": "^0.12.1",
@@ -20,7 +21,8 @@
     "js-yaml": "^4.1.0",
     "jsonpath": "^1.1.1",
     "lzutf8": "^0.6.0",
-    "talisman": "^1.1.4"
+    "talisman": "^1.1.4",
+    "yargs": "^17.2.1"
   },
   "devDependencies": {
     "chai": "^4.3.4",
@@ -33,7 +35,6 @@
     "gulp-transform": "^3.0.5",
     "mocha": "^9.1.3",
     "nyc": "^15.1.0",
-    "sinon": "^12.0.1",
-    "yargs": "^17.2.1"
+    "sinon": "^12.0.1"
   }
 }
