+++ test/data/repo/github/repository-response.json
@@ -1,0 +1,142 @@
+{
+  "id": 334182395,
+  "node_id": "MDEwOlJlcG9zaXRvcnkzMzQxODIzOTU=",
+  "name": "openapi-chopper",
+  "full_name": "api-stuff/openapi-chopper",
+  "private": false,
+  "owner": {
+    "login": "api-stuff",
+    "id": 68026188,
+    "node_id": "MDEyOk9yZ2FuaXphdGlvbjY4MDI2MTg4",
+    "avatar_url": "https://avatars.githubusercontent.com/u/68026188?v=4",
+    "gravatar_id": "",
+    "url": "https://api.github.com/users/api-stuff",
+    "html_url": "https://github.com/api-stuff",
+    "followers_url": "https://api.github.com/users/api-stuff/followers",
+    "following_url": "https://api.github.com/users/api-stuff/following{/other_user}",
+    "gists_url": "https://api.github.com/users/api-stuff/gists{/gist_id}",
+    "starred_url": "https://api.github.com/users/api-stuff/starred{/owner}{/repo}",
+    "subscriptions_url": "https://api.github.com/users/api-stuff/subscriptions",
+    "organizations_url": "https://api.github.com/users/api-stuff/orgs",
+    "repos_url": "https://api.github.com/users/api-stuff/repos",
+    "events_url": "https://api.github.com/users/api-stuff/events{/privacy}",
+    "received_events_url": "https://api.github.com/users/api-stuff/received_events",
+    "type": "Organization",
+    "site_admin": false
+  },
+  "html_url": "https://github.com/api-stuff/openapi-chopper",
+  "description": "For chopping up OpenAPI specification documents... like it says on the tin...",
+  "fork": false,
+  "url": "https://api.github.com/repos/api-stuff/openapi-chopper",
+  "forks_url": "https://api.github.com/repos/api-stuff/openapi-chopper/forks",
+  "keys_url": "https://api.github.com/repos/api-stuff/openapi-chopper/keys{/key_id}",
+  "collaborators_url": "https://api.github.com/repos/api-stuff/openapi-chopper/collaborators{/collaborator}",
+  "teams_url": "https://api.github.com/repos/api-stuff/openapi-chopper/teams",
+  "hooks_url": "https://api.github.com/repos/api-stuff/openapi-chopper/hooks",
+  "issue_events_url": "https://api.github.com/repos/api-stuff/openapi-chopper/issues/events{/number}",
+  "events_url": "https://api.github.com/repos/api-stuff/openapi-chopper/events",
+  "assignees_url": "https://api.github.com/repos/api-stuff/openapi-chopper/assignees{/user}",
+  "branches_url": "https://api.github.com/repos/api-stuff/openapi-chopper/branches{/branch}",
+  "tags_url": "https://api.github.com/repos/api-stuff/openapi-chopper/tags",
+  "blobs_url": "https://api.github.com/repos/api-stuff/openapi-chopper/git/blobs{/sha}",
+  "git_tags_url": "https://api.github.com/repos/api-stuff/openapi-chopper/git/tags{/sha}",
+  "git_refs_url": "https://api.github.com/repos/api-stuff/openapi-chopper/git/refs{/sha}",
+  "trees_url": "https://api.github.com/repos/api-stuff/openapi-chopper/git/trees{/sha}",
+  "statuses_url": "https://api.github.com/repos/api-stuff/openapi-chopper/statuses/{sha}",
+  "languages_url": "https://api.github.com/repos/api-stuff/openapi-chopper/languages",
+  "stargazers_url": "https://api.github.com/repos/api-stuff/openapi-chopper/stargazers",
+  "contributors_url": "https://api.github.com/repos/api-stuff/openapi-chopper/contributors",
+  "subscribers_url": "https://api.github.com/repos/api-stuff/openapi-chopper/subscribers",
+  "subscription_url": "https://api.github.com/repos/api-stuff/openapi-chopper/subscription",
+  "commits_url": "https://api.github.com/repos/api-stuff/openapi-chopper/commits{/sha}",
+  "git_commits_url": "https://api.github.com/repos/api-stuff/openapi-chopper/git/commits{/sha}",
+  "comments_url": "https://api.github.com/repos/api-stuff/openapi-chopper/comments{/number}",
+  "issue_comment_url": "https://api.github.com/repos/api-stuff/openapi-chopper/issues/comments{/number}",
+  "contents_url": "https://api.github.com/repos/api-stuff/openapi-chopper/contents/{+path}",
+  "compare_url": "https://api.github.com/repos/api-stuff/openapi-chopper/compare/{base}...{head}",
+  "merges_url": "https://api.github.com/repos/api-stuff/openapi-chopper/merges",
+  "archive_url": "https://api.github.com/repos/api-stuff/openapi-chopper/{archive_format}{/ref}",
+  "downloads_url": "https://api.github.com/repos/api-stuff/openapi-chopper/downloads",
+  "issues_url": "https://api.github.com/repos/api-stuff/openapi-chopper/issues{/number}",
+  "pulls_url": "https://api.github.com/repos/api-stuff/openapi-chopper/pulls{/number}",
+  "milestones_url": "https://api.github.com/repos/api-stuff/openapi-chopper/milestones{/number}",
+  "notifications_url": "https://api.github.com/repos/api-stuff/openapi-chopper/notifications{?since,all,participating}",
+  "labels_url": "https://api.github.com/repos/api-stuff/openapi-chopper/labels{/name}",
+  "releases_url": "https://api.github.com/repos/api-stuff/openapi-chopper/releases{/id}",
+  "deployments_url": "https://api.github.com/repos/api-stuff/openapi-chopper/deployments",
+  "created_at": "2021-01-29T15:19:21Z",
+  "updated_at": "2021-12-01T07:56:59Z",
+  "pushed_at": "2021-02-09T09:54:42Z",
+  "git_url": "git://github.com/api-stuff/openapi-chopper.git",
+  "ssh_url": "git@github.com:api-stuff/openapi-chopper.git",
+  "clone_url": "https://github.com/api-stuff/openapi-chopper.git",
+  "svn_url": "https://github.com/api-stuff/openapi-chopper",
+  "homepage": "",
+  "size": 65,
+  "stargazers_count": 0,
+  "watchers_count": 0,
+  "language": "JavaScript",
+  "has_issues": true,
+  "has_projects": true,
+  "has_downloads": true,
+  "has_wiki": true,
+  "has_pages": false,
+  "forks_count": 0,
+  "mirror_url": null,
+  "archived": false,
+  "disabled": false,
+  "open_issues_count": 0,
+  "license": {
+    "key": "mit",
+    "name": "MIT License",
+    "spdx_id": "MIT",
+    "url": "https://api.github.com/licenses/mit",
+    "node_id": "MDc6TGljZW5zZTEz"
+  },
+  "allow_forking": true,
+  "is_template": false,
+  "topics": [
+    "openapi"
+  ],
+  "visibility": "public",
+  "forks": 0,
+  "open_issues": 0,
+  "watchers": 0,
+  "default_branch": "master",
+  "permissions": {
+    "admin": true,
+    "maintain": true,
+    "push": true,
+    "triage": true,
+    "pull": true
+  },
+  "temp_clone_token": "",
+  "allow_squash_merge": true,
+  "allow_merge_commit": true,
+  "allow_rebase_merge": true,
+  "allow_auto_merge": false,
+  "delete_branch_on_merge": false,
+  "allow_update_branch": false,
+  "organization": {
+    "login": "api-stuff",
+    "id": 68026188,
+    "node_id": "MDEyOk9yZ2FuaXphdGlvbjY4MDI2MTg4",
+    "avatar_url": "https://avatars.githubusercontent.com/u/68026188?v=4",
+    "gravatar_id": "",
+    "url": "https://api.github.com/users/api-stuff",
+    "html_url": "https://github.com/api-stuff",
+    "followers_url": "https://api.github.com/users/api-stuff/followers",
+    "following_url": "https://api.github.com/users/api-stuff/following{/other_user}",
+    "gists_url": "https://api.github.com/users/api-stuff/gists{/gist_id}",
+    "starred_url": "https://api.github.com/users/api-stuff/starred{/owner}{/repo}",
+    "subscriptions_url": "https://api.github.com/users/api-stuff/subscriptions",
+    "organizations_url": "https://api.github.com/users/api-stuff/orgs",
+    "repos_url": "https://api.github.com/users/api-stuff/repos",
+    "events_url": "https://api.github.com/users/api-stuff/events{/privacy}",
+    "received_events_url": "https://api.github.com/users/api-stuff/received_events",
+    "type": "Organization",
+    "site_admin": false
+  },
+  "network_count": 0,
+  "subscribers_count": 2
+}
